+++ sdk-okhttp-client/src/test/java/io/nem/sdk/infrastructure/okhttp/AbstractOkHttpRespositoryTest.java
@@ -25,7 +25,8 @@ import io.nem.sdk.openapi.okhttp_gson.invoker.ApiClient;
 import io.nem.sdk.openapi.okhttp_gson.invoker.ApiException;
 import io.nem.sdk.openapi.okhttp_gson.invoker.ApiResponse;
 import io.nem.sdk.openapi.okhttp_gson.invoker.JSON;
-import io.nem.sdk.openapi.okhttp_gson.model.NodeInfoDTO;
+import io.nem.sdk.openapi.okhttp_gson.model.NetworkTypeDTO;
+import io.nem.sdk.openapi.okhttp_gson.model.NetworkTypeNameEnum;
 import java.lang.reflect.Type;
 import java.util.Collections;
 import java.util.HashMap;
@@ -118,8 +119,8 @@ public abstract class AbstractOkHttpRespositoryTest {
 
 
     protected final NetworkType resolveNetworkType() throws ApiException {
-        NodeInfoDTO networkTypeDTO = new NodeInfoDTO();
-        networkTypeDTO.setNetworkIdentifier(NetworkType.MIJIN_TEST.getValue());
+        NetworkTypeDTO networkTypeDTO = new NetworkTypeDTO();
+        networkTypeDTO.setName(NetworkTypeNameEnum.MIJINTEST);
         mockRemoteCall(networkTypeDTO);
         NetworkType networkType = getRepository().getNetworkTypeBlocking();
         Assertions.assertEquals(NetworkType.MIJIN_TEST, networkType);