+++ travis/build.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -e
+
+echo "Building branch '$TRAVIS_BRANCH'"
+./gradlew install jacocoTestReport coveralls -x javadoc